+++ src/usb.c
@@ -1,49 +1,49 @@
 
 
-//#include "everdrive.h"
-//#include "sys.h"
-//#include "types.h"
-//#include <libdragon.h>
-//#include "rom.h"
+#include "everdrive.h"
+#include "sys.h"
+#include "types.h"
+#include <libdragon.h>
+#include "rom.h"
 
-//u8 cmdTest();
-//u8 cmdFill();
-//u8 cmdReadRom();
-//u8 cmdWriteRom();
+u8 cmdTest();
+u8 cmdFill();
+u8 cmdReadRom();
+u8 cmdWriteRom();
 
-//u64 usb_buff[128];
-//u8 *usb_buff8; // = (u8 *) usb_buff;
+u64 usb_buff[128];
+u8 *usb_buff8; // = (u8 *) usb_buff;
 
 
-//#define PI_BSD_DOM1_LAT_REG	(PI_BASE_REG+0x14)
+#define PI_BSD_DOM1_LAT_REG	(PI_BASE_REG+0x14)
 
 /* PI dom1 pulse width (R/W): [7:0] domain 1 device R/W strobe pulse width */
-//#define PI_BSD_DOM1_PWD_REG	(PI_BASE_REG+0x18)
+#define PI_BSD_DOM1_PWD_REG	(PI_BASE_REG+0x18)
 
 /* PI dom1 page size (R/W): [3:0] domain 1 device page size */
-//#define PI_BSD_DOM1_PGS_REG	(PI_BASE_REG+0x1C)    /*   page size */
+#define PI_BSD_DOM1_PGS_REG	(PI_BASE_REG+0x1C)    /*   page size */
 
 /* PI dom1 release (R/W): [1:0] domain 1 device R/W release duration */
-//#define PI_BSD_DOM1_RLS_REG	(PI_BASE_REG+0x20)
+#define PI_BSD_DOM1_RLS_REG	(PI_BASE_REG+0x20)
 /* PI dom2 latency (R/W): [7:0] domain 2 device latency */
-//#define PI_BSD_DOM2_LAT_REG	(PI_BASE_REG+0x24)    /* Domain 2 latency */
+#define PI_BSD_DOM2_LAT_REG	(PI_BASE_REG+0x24)    /* Domain 2 latency */
 
 /* PI dom2 pulse width (R/W): [7:0] domain 2 device R/W strobe pulse width */
-//#define PI_BSD_DOM2_PWD_REG	(PI_BASE_REG+0x28)    /*   pulse width */
+#define PI_BSD_DOM2_PWD_REG	(PI_BASE_REG+0x28)    /*   pulse width */
 
 /* PI dom2 page size (R/W): [3:0] domain 2 device page size */
-//#define PI_BSD_DOM2_PGS_REG	(PI_BASE_REG+0x2C)    /*   page size */
+#define PI_BSD_DOM2_PGS_REG	(PI_BASE_REG+0x2C)    /*   page size */
 
 /* PI dom2 release (R/W): [1:0] domain 2 device R/W release duration */
-//#define PI_BSD_DOM2_RLS_REG	(PI_BASE_REG+0x30)    /*   release duration */
+#define PI_BSD_DOM2_RLS_REG	(PI_BASE_REG+0x30)    /*   release duration */
 
-//#define	PHYS_TO_K1(x)	((u32)(x)|0xA0000000)	/* physical to kseg1 */
-//#define	IO_WRITE(addr,data)	(*(volatile u32*)PHYS_TO_K1(addr)=(u32)(data))
-//#define PI_BASE_REG		0x04600000
+#define	PHYS_TO_K1(x)	((u32)(x)|0xA0000000)	/* physical to kseg1 */
+#define	IO_WRITE(addr,data)	(*(volatile u32*)PHYS_TO_K1(addr)=(u32)(data))
+#define PI_BASE_REG		0x04600000
 
-//extern u8 system_cic;
+extern u8 system_cic;
 
-/*u8 usbListener() {
+u8 usbListener() {
 
     volatile u16 resp;
     volatile u8 cmd;
@@ -83,12 +83,17 @@
             //IO_WRITE(PI_BSD_DOM1_PGS_REG, 0x0c);
             //IO_WRITE(PI_BSD_DOM1_PGS_REG, 0x80);
             //evdSetESaveType(SAVE_TYPE_EEP16k);
-            system_cic = CIC_6102;
+            //system_cic = CIC_6102; //TODO: re-enable
             evd_lockRegs();
             IO_WRITE(PI_STATUS_REG, 3);
             sleep(2);
             pif_boot();
             break;
+        case 'D':
+            //TODO: initiate debug session                    
+        break;
+        default:
+        break;
 
     }
 
@@ -117,7 +122,7 @@ u8 cmdFill() {
         usb_buff8[i] = 0;
     }
     //console_printf("buff prepared\n");
-    romFill(0, 0x200000, 0);
+    //romFill(0, 0x200000, 0); //TODO: re-enable
     //console_printf("fill done\n");
 
     usb_buff8[3] = 'k';
@@ -178,4 +183,3 @@ u8 cmdWriteRom() {
     return 0;
 
 }
-*/
